+++ addon/-legacy-private/system/model/states.js
@@ -521,6 +521,7 @@ const RootState = {
   loading: {
     // FLAGS
     isLoading: true,
+    isEmpty: true,
 
     exit(internalModel) {
       internalModel._promiseProxy = null;