+++ addon/-legacy-private/system/relationships/state/belongs-to.js
@@ -34,6 +34,7 @@ export default class BelongsToRelationship extends Relationship {
     } else if (this.inverseInternalModel) {
       this.removeInternalModel(this.inverseInternalModel);
     }
+
     this.setHasAnyRelationshipData(true);
     this.setRelationshipIsStale(false);
     this.setRelationshipIsEmpty(false);
@@ -72,12 +73,14 @@ export default class BelongsToRelationship extends Relationship {
     }
 
     this.canonicalState = internalModel;
+    this.setHasAnyRelationshipData(true);
+    this.setRelationshipIsEmpty(false);
     super.addCanonicalInternalModel(internalModel);
   }
 
   inverseDidDematerialize() {
     super.inverseDidDematerialize(this.inverseInternalModel);
-    this.notifyBelongsToChanged();
+    this.notifyBelongsToChange();
   }
 
   removeCompletelyFromOwn(internalModel) {
@@ -89,7 +92,7 @@ export default class BelongsToRelationship extends Relationship {
 
     if (this.inverseInternalModel === internalModel) {
       this.inverseInternalModel = null;
-      this.notifyBelongsToChanged();
+      this.notifyBelongsToChange();
     }
   }
 
@@ -108,7 +111,7 @@ export default class BelongsToRelationship extends Relationship {
     if (this.inverseInternalModel !== this.canonicalState) {
       this.inverseInternalModel = this.canonicalState;
       this._promiseProxy = null;
-      this.notifyBelongsToChanged();
+      this.notifyBelongsToChange();
     }
 
     super.flushCanonical();
@@ -127,7 +130,7 @@ export default class BelongsToRelationship extends Relationship {
 
     this.inverseInternalModel = internalModel;
     super.addInternalModel(internalModel);
-    this.notifyBelongsToChanged();
+    this.notifyBelongsToChange();
   }
 
   setRecordPromise(belongsToPromise) {
@@ -150,18 +153,24 @@ export default class BelongsToRelationship extends Relationship {
     this.inverseInternalModel = null;
     this._promiseProxy = null;
     super.removeInternalModelFromOwn(internalModel);
-    this.notifyBelongsToChanged();
+    this.notifyBelongsToChange();
   }
 
   removeAllInternalModelsFromOwn() {
     super.removeAllInternalModelsFromOwn();
     this.inverseInternalModel = null;
     this._promiseProxy = null;
-    this.notifyBelongsToChanged();
+    this.notifyBelongsToChange();
   }
 
-  notifyBelongsToChanged() {
-    this.internalModel.notifyBelongsToChanged(this.key);
+  notifyBelongsToChange() {
+    if (this._promiseProxy !== null) {
+      let iM = this.inverseInternalModel;
+
+      this._updateLoadingPromise(proxyRecord(iM), iM ? iM.getRecord() : null);
+    }
+
+    this.internalModel.notifyBelongsToChange(this.key);
   }
 
   removeCanonicalInternalModelFromOwn(internalModel) {
@@ -169,6 +178,8 @@ export default class BelongsToRelationship extends Relationship {
       return;
     }
     this.canonicalState = null;
+    this.setHasAnyRelationshipData(true);
+    this.setRelationshipIsEmpty(true);
     super.removeCanonicalInternalModelFromOwn(internalModel);
   }
 
@@ -238,9 +249,7 @@ export default class BelongsToRelationship extends Relationship {
 
     if (this.isAsync) {
       if (this._promiseProxy === null) {
-        let promise = resolve(this.inverseInternalModel).then(internalModel => {
-          return internalModel ? internalModel.getRecord() : null;
-        });
+        let promise = proxyRecord(this.inverseInternalModel);
         this._updateLoadingPromise(promise, record);
       }
 
@@ -276,6 +285,12 @@ export default class BelongsToRelationship extends Relationship {
       this.setCanonicalInternalModel(internalModel);
     }
   }
+}
+
+function proxyRecord(internalModel) {
+  return resolve(internalModel).then(resolvedInternalModel => {
+    return resolvedInternalModel ? resolvedInternalModel.getRecord() : null;
+  });
 }
 
 function handleCompletedFind(relationship, error) {