+++ addon/-record-data-private/system/model/states.js
@@ -518,6 +518,7 @@ const RootState = {
   // XHR to retrieve the data.
   loading: {
     // FLAGS
+    isEmpty: true,
     isLoading: true,
 
     exit(internalModel) {