+++ addon/-record-data-private/system/relationships/state/relationship.js
@@ -55,6 +55,7 @@ export default class Relationship {
   constructor(store, inverseKey, relationshipMeta, modelData, inverseIsAsync) {
     heimdall.increment(newRelationship);
     this.inverseIsAsync = inverseIsAsync;
+    this.kind = relationshipMeta.kind;
     let async = relationshipMeta.options.async;
     let polymorphic = relationshipMeta.options.polymorphic;
     this.modelData = modelData;
@@ -73,18 +74,39 @@ export default class Relationship {
     this.__inverseMeta = undefined;
 
     /*
+     This flag forces fetch. `true` for a single request once `reload()`
+       has been called `false` at all other times.
+    */
+    this.shouldForceReload = false;
+
+    /*
        This flag indicates whether we should
         re-fetch the relationship the next time
         it is accessed.
 
+        The difference between this flag and `shouldForceReload`
+        is in how we treat the presence of partially missing data:
+          - for a forced reload, we will reload the link or EVERY record
+          - for a stale reload, we will reload the link (if present) else only MISSING records
+
+        Ideally these flags could be merged, but because we don't give the
+        request layer the option of deciding how to resolve the data being queried
+        we are forced to differentiate for now.
+
+        It is also possible for a relationship to remain stale after a forced reload; however,
+        in this case `hasFailedLoadAttempt` ought to be `true`.
+
       false when
-        => initial setup
+        => modelData.isNew() on initial setup
         => a previously triggered request has resolved
         => we get relationship data via push
 
       true when
-        => relationship.reload() has been called
+        => !modelData.isNew() on initial setup
+        => an inverse has been unloaded
         => we get a new link for the relationship
+
+      TODO @runspired unskip the acceptance tests and fix these flags
      */
     this.relationshipIsStale = false;
 
@@ -149,7 +171,7 @@ export default class Relationship {
     this.relationshipIsEmpty = true;
 
     /*
-      Flag def here for reference, defined as getter below
+      Flag def here for reference, defined as getter in has-many.js / belongs-to.js
 
       true when
         => hasAnyRelationshipData is true
@@ -158,7 +180,6 @@ export default class Relationship {
 
       TODO, consider changing the conditional here from !isEmpty to !hiddenFromRecordArrays
     */
-    // this.allInverseRecordsAreLoaded = false;
 
     // TODO do we want this anymore? Seems somewhat useful
     //   especially if we rename to `hasUpdatedLink`
@@ -167,8 +188,8 @@ export default class Relationship {
     // this.updatedLink = false;
   }
 
-  get allInverseRecordsAreLoaded() {
-    return !this.localStateIsEmpty();
+  get isNew() {
+    return this.modelData.isNew();
   }
 
   _inverseIsAsync() {
@@ -212,6 +233,9 @@ export default class Relationship {
     if (!this.inverseKey) {
       return;
     }
+    // TODO @runspired fairly sure we need to become stale here
+    // this.setRelationshipIsStale(true);
+
     // we actually want a union of members and canonicalMembers
     // they should be disjoint but currently are not due to a bug
     this.forAllMembers(inverseModelData => {
@@ -280,6 +304,7 @@ export default class Relationship {
   }
 
   removeAllModelDatasFromOwn() {
+    this.setRelationshipIsStale(true);
     this.members.clear();
   }
 
@@ -548,7 +573,7 @@ export default class Relationship {
     warn(
       `You pushed a record of type '${this.modelData.modelName}' with a relationship '${
         this.key
-      }' configured as 'async: false'. You've included a link but no primary data, this may be an error in your payload.`,
+      }' configured as 'async: false'. You've included a link but no primary data, this may be an error in your payload. EmberData will treat this relationship as known-to-be-empty.`,
       this.isAsync || this.hasAnyRelationshipData,
       {
         id: 'ds.store.push-link-for-sync-relationship',
@@ -623,6 +648,11 @@ export default class Relationship {
     if (payload.data !== undefined) {
       hasRelationshipDataProperty = true;
       this.updateData(payload.data, initial);
+    } else if (this.isAsync === false) {
+      hasRelationshipDataProperty = true;
+      let data = this.kind === 'hasMany' ? [] : null;
+
+      this.updateData(data, initial);
     }
 
     if (payload.links && payload.links.related) {