+++ tests/integration/relationships/one-to-one-test.js
@@ -21,6 +21,7 @@ module('integration/relationships/one_to_one_test - OneToOne relationships', {
     });
 
     Job = DS.Model.extend({
+      name: attr(),
       isGood: attr(),
       user: belongsTo('user', { async: false }),
     });
@@ -203,137 +204,194 @@ test('Fetching a belongsTo that is set to null removes the record from a relatio
   assert.equal(job.get('user'), null, 'User relationship was removed correctly');
 });
 
-test('Fetching a belongsTo that is set to a different record, sets the old relationship to null - async', function(assert) {
-  assert.expect(3);
-  var stanley, stanleysFriend;
-  run(function() {
-    stanley = store.push({
-      data: {
-        id: 1,
-        type: 'user',
-        attributes: {
-          name: 'Stanley',
+test('Fetching a belongsTo that is set to a different record, sets the old relationship to null - async', async function(assert) {
+  let user1 = store.push({
+    data: {
+      type: 'user',
+      id: '1',
+      attributes: { name: 'Chris' },
+      relationships: {
+        bestFriend: {
+          data: { type: 'user', id: '2' },
         },
+      },
+    },
+    included: [
+      {
+        type: 'user',
+        id: '2',
+        attributes: { name: 'Igor' },
         relationships: {
           bestFriend: {
-            data: {
-              id: 2,
-              type: 'user',
-            },
+            data: { type: 'user', id: '1' },
           },
         },
       },
-    });
-    stanleysFriend = store.push({
-      data: {
-        id: 2,
-        type: 'user',
-        attributes: {
-          name: "Stanley's friend",
-        },
-        relationships: {
-          bestFriend: {
-            data: {
-              id: 1,
-              type: 'user',
-            },
-          },
-        },
-      },
-    });
-
-    stanleysFriend.get('bestFriend').then(function(fetchedUser) {
-      assert.equal(fetchedUser, stanley, 'User relationship was initally setup correctly');
-      var stanleysNewFriend;
-      run(function() {
-        stanleysNewFriend = store.push({
-          data: {
-            id: 3,
-            type: 'user',
-            attributes: {
-              name: "Stanley's New friend",
-            },
-            relationships: {
-              bestFriend: {
-                data: {
-                  id: 1,
-                  type: 'user',
-                },
-              },
-            },
-          },
-        });
-      });
-
-      stanley.get('bestFriend').then(function(fetchedNewFriend) {
-        assert.equal(
-          fetchedNewFriend,
-          stanleysNewFriend,
-          'User relationship was updated correctly'
-        );
-      });
-
-      stanleysFriend.get('bestFriend').then(function(fetchedOldFriend) {
-        assert.equal(fetchedOldFriend, null, 'The old relationship was set to null correctly');
-      });
-    });
+    ],
   });
+
+  let user2 = store.peekRecord('user', '2');
+  let user1Friend = await user1.get('bestFriend');
+
+  assert.equal(user1Friend, user2, '<user:1>.bestFriend is <user:2>');
+
+  /*
+    Now we "reload" <user:2> but with a new bestFriend. While this only gives
+    us new canonical information for <user:2> and <user:3>, it also severs
+    the previous canonical relationship with <user:1>. We infer from this
+    that the new canonical state for <user:1>.bestFriend is `null`.
+
+    Users for whom this is not true should either
+
+    - include information for user:1 in the payload severing this link
+    - manually reload user:1 or use the belongsToReference to reload user:1.bestFriend
+   */
+  store.push({
+    data: {
+      type: 'user',
+      id: '2',
+      attributes: { name: 'Igor' },
+      relationships: {
+        bestFriend: {
+          data: { type: 'user', id: '3' },
+        },
+      },
+    },
+    included: [
+      {
+        type: 'user',
+        id: '3',
+        attributes: { name: 'Evan' },
+        relationships: {
+          bestFriend: {
+            data: { type: 'user', id: '2' },
+          },
+        },
+      },
+    ],
+  });
+
+  let user3 = store.peekRecord('user', '3');
+  let user1bestFriend = await user1.get('bestFriend');
+  let user2bestFriend = await user2.get('bestFriend');
+  let user3bestFriend = await user3.get('bestFriend');
+
+  assert.equal(user3bestFriend, user2, '<user:3>.bestFriend is <user:2>');
+  assert.equal(user2bestFriend, user3, '<user:2>.bestFriend is <user:3>');
+  assert.equal(user1bestFriend, null, '<user:1>.bestFriend is null');
+
+  let user1bestFriendState = user1.belongsTo('bestFriend').belongsToRelationship;
+
+  assert.equal(user1bestFriendState.canonicalState, null, '<user:1>.job is canonically empty');
+  assert.equal(user1bestFriendState.currentState, null, '<user:1>.job is locally empty');
+  assert.equal(user1bestFriendState.relationshipIsEmpty, true, 'The relationship is empty');
+  assert.equal(user1bestFriendState.relationshipIsStale, false, 'The relationship is not stale');
+  assert.equal(
+    user1bestFriendState.shouldForceReload,
+    false,
+    'The relationship does not require reload'
+  );
+  assert.equal(
+    user1bestFriendState.hasAnyRelationshipData,
+    true,
+    'The relationship considers its canonical data complete'
+  );
+  assert.equal(
+    user1bestFriendState.allInverseRecordsAreLoaded,
+    true,
+    'The relationship has all required data'
+  );
 });
 
-test('Fetching a belongsTo that is set to a different record, sets the old relationship to null - sync', function(assert) {
-  var job, user, newBetterJob;
-  run(function() {
-    job = store.push({
-      data: {
-        id: 2,
-        type: 'job',
-        attributes: {
-          isGood: false,
+test('Fetching a belongsTo that is set to a different record, sets the old relationship to null - sync', async function(assert) {
+  let user1 = store.push({
+    data: {
+      type: 'user',
+      id: '1',
+      attributes: { name: 'Chris' },
+      relationships: {
+        job: {
+          data: { type: 'job', id: '1' },
         },
       },
-    });
-    user = store.push({
-      data: {
-        id: 1,
-        type: 'user',
-        attributes: {
-          name: 'Stanley',
-        },
-        relationships: {
-          job: {
-            data: {
-              id: 2,
-              type: 'job',
-            },
-          },
-        },
-      },
-    });
-  });
-  assert.equal(job.get('user'), user, 'Job and user initially setup correctly');
-  run(function() {
-    newBetterJob = store.push({
-      data: {
-        id: 3,
+    },
+    included: [
+      {
         type: 'job',
-        attributes: {
-          isGood: true,
-        },
+        id: '1',
+        attributes: { name: 'Golf Picker Mechanic' },
         relationships: {
           user: {
-            data: {
-              id: 1,
-              type: 'user',
-            },
+            data: { type: 'user', id: '1' },
           },
         },
       },
-    });
+    ],
   });
 
-  assert.equal(user.get('job'), newBetterJob, 'Job updated correctly');
-  assert.equal(job.get('user'), null, 'Old relationship nulled out correctly');
-  assert.equal(newBetterJob.get('user'), user, 'New job setup correctly');
+  let job1 = store.peekRecord('job', '1');
+
+  assert.equal(user1.get('job'), job1, '<user:1>.job is <job:1>');
+
+  /*
+    Now we "reload" <job:1> but with a new user. While this only gives
+    us new canonical information for <job:1> and <user:2>, it also severs
+    the previous canonical relationship with <user:1>. We infer from this
+    that the new canonical state for <user:1>.job is `null`.
+
+    Users for whom this is not true should either
+
+    - include information for user:1 in the payload severing this link
+    - manually reload user:1 or use the belongsToReference to reload user:1.job
+   */
+  store.push({
+    data: {
+      type: 'job',
+      id: '1',
+      attributes: { name: 'Golf Picker Mechanic' },
+      relationships: {
+        user: {
+          data: { type: 'user', id: '2' },
+        },
+      },
+    },
+    included: [
+      {
+        type: 'user',
+        id: '2',
+        attributes: { name: 'Evan' },
+        relationships: {
+          job: {
+            data: { type: 'job', id: '1' },
+          },
+        },
+      },
+    ],
+  });
+
+  let user2 = store.peekRecord('user', '2');
+
+  assert.equal(user2.get('job'), job1, '<user:2>.job is <job:1>');
+  assert.equal(job1.get('user'), user2, '<job:1>.user is <user:2>');
+  assert.equal(user1.get('job'), null, '<user:1>.job is null');
+
+  let user1JobState = user1.belongsTo('job').belongsToRelationship;
+
+  assert.equal(user1JobState.canonicalState, null, '<user:1>.job is canonically empty');
+  assert.equal(user1JobState.currentState, null, '<user:1>.job is locally empty');
+  assert.equal(user1JobState.relationshipIsEmpty, true, 'The relationship is empty');
+  assert.equal(user1JobState.relationshipIsStale, false, 'The relationship is not stale');
+  assert.equal(user1JobState.shouldForceReload, false, 'The relationship does not require reload');
+  assert.equal(
+    user1JobState.hasAnyRelationshipData,
+    true,
+    'The relationship considers its canonical data complete'
+  );
+  assert.equal(
+    user1JobState.allInverseRecordsAreLoaded,
+    true,
+    'The relationship has all required data'
+  );
 });
 
 /*